requirements.txt
# ==============================
# 🔥 DJANGO CORE
# ==============================
Django
djangorestframework

# ==============================
# 🌍 CORS (React <-> Django)
# ==============================
django-cors-headers

# ==============================
# 🔐 ENV VARIABLES
# ==============================
python-dotenv

# ==============================
# 🗺️ GEO (distance / localisation)
# ==============================
geopy>=2.4

# ==============================
# 🤖 OPENAI (IA)
# ==============================
openai>=1.0

# ==============================
# 🖼️ IMAGE UPLOAD
# ==============================
Pillow>=10.0

# ==============================
# ⚡ UTILITAIRES
# ==============================
requests>=2.31

# ==============================
# 🧪 TESTS (optionnel mais recommandé)
# ==============================
pytest>=8.0
pytest-django>=4.7

# ==============================
# 🐍 SERVER (PROD READY LATER)
# ==============================
gunicorn>=21.2